Fix implementation of thrust::detail::uint64_t on 32b platforms.
--HG--
extra : convert_revision : svn%3A83215879-3e5a-4751-8c9d-778f44bb06a5/trunk%40841

+++ thrust/detail/cstdint.h
@@ -63,7 +63,7 @@ template<int word_size = sizeof(int*)> struct divine_uint64_t;
 template<>
   struct divine_uint64_t<4>
 {
-  typedef unsigned long int type;
+  typedef unsigned long long int type;
 };
 
 // 64b machine type
